+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/room/RoomPicture.kt
@@ -14,4 +14,5 @@ import androidx.room.*
 )
 data class RoomPicture(
     @PrimaryKey val uri: Uri,
+    val categoryId: String
 )
